Split panel and arakawa reports creation in different scripts.
